Another attempt for PNET and updated README
This attempt tries to modify the patch directly, i.e. the patch
itself is watch by tensorflow and a gradient for the patch is
created.

Sadly the gradient isn't working again, since his entries are
very small. Sometimes the gradient only contains 0. Another
problem is that the gradient only changes very small parts.