Ported more code to Tensorflow and changed how resize is used
math. was changed to math from Tensorflow and removed.

All instances of tf.image.resize() use now method='lanczos5' and
antialias=True.

Instead of our Non Maximum Suppression method I use the NMS method
given by TensorFlow. And removed a mistake in the second nms instance.
It used tf_pick instead of the total_boxes.

The variables bounding_boxes and image are now a Tensors
before they are used in the Gradient_Tape enviroment.
I hoped that my gradient wasn't working properly since he does not see
these to variables. But this was not the case.