Solution against the memory problem + addition of convenience
Found a solution against the usage of too much memory. The problem was that
MTCNN allocated memory for itself (especially the executions of the
functions of its model ONET) and never cleared the allocations. Now, at the beginning of each
beginning of each epoch we generate one new MTCNN, so the older MTCNN
is cleared and therefore the memory used by its models PNET, RNET and
ONET is also cleared.

In addidtion I added some convenience into the code:
Now, we return images only at the first epoch and every 20th epoch.
Also, now we also return the patch separately.