Structural changes (Only one Patch for all pictures) + Can read different directories
The structure of test2.py was changes, so that there is only one patch
for multiple images, like it was stated in the paper.

At the bottom of the code, you can choose to execute three different versions:
1. picture_images(): Uses the images from the github adversial-attack-on-mtnn.
2. face_Control_all(): Uses all the images in Face_control/0_Parade. ATTENTION
there are 460 images therefore this execution takes a lot of time for
each epoch.
3. face_control_variable(): Uses a variable number of images from
Face_control/0_Parade. ATTENTION: There is a problem with pCount == 66.

Right now, I am working on the implementation of IoU into the tape, since
the paper needs it before the loss is calculated.
And I am also trying to fix the error of face_control_variable().