Added new run file and fixed MTCNN
In __my_tf_apply_patch only applied the patch to the last face in an image. Now it
applies the patch on all faces in an image.

test_on_server.py is similar to test.py but it doesn't create ASes. And
therefore it outputs only the patch and can't output the images with the
applied patch.