Improved the the python files for patch testing
Improved face_detection.py and face_test.py in the
sense that they use less memory, retunrn also the
confidence score in the txt file, show the highest
given confidence score of an image and round the
confidence score to the important digits. Now,
they return the attacked images without the need to
create an ASes.