Added user friendly versions of test_on_server.py and face_test.py
friendly_test_on_server.py is the user friendly version of test_on_server.py  and  friendly_face_test.py is the user friendly version of face_test.py.

They are working with the folder  "User_Friendly", which is divided into the folders "Train" and "Test". "Train" is used by friendly_test_on_server.py and "Test" is used by friendly_face_test.py.

Both versions are user friendly because the user does not need to prepare ground truths for each image but only has to add images into the folder "Images". The used ground truths will be then the found faces by original_MTCNN before the images are changed.

original_mtcnn.py is the original code of MTCNN and used to create the ground truths and test the patches.